Add isAdmin and Auth

+++ routers/categories.js
@@ -1,8 +1,13 @@
 const { Category } = require("../models/categories");
 const express = require("express");
+const auth = require("../auth/auth");
 const router = express.Router();
 
 router.get(`/`, async (req, res) => {
+  const authResult = await auth(req, res, true);
+  if (!authResult)
+    return res.status(500).json({ success: false, err: "auth err" });
+
   categoryList = await Category.find();
 
   if (!categoryList) {
@@ -11,6 +16,10 @@ router.get(`/`, async (req, res) => {
   res.send(categoryList);
 });
 router.get("/:id", async (req, res) => {
+  const authResult = await auth(req, res, true);
+  if (!authResult)
+    return res.status(500).json({ success: false, err: "auth err" });
+
   const category = await Category.findById(req.params.id);
   if (!category) {
     res.status(500).json({
@@ -20,6 +29,10 @@ router.get("/:id", async (req, res) => {
   res.status(200).send(category);
 });
 router.post("/", async (req, res) => {
+  const authResult = await auth(req, res, true);
+  if (!authResult)
+    return res.status(500).json({ success: false, err: "auth err" });
+
   let newCategory = new Category();
   newCategory.name = req.body.name;
   newCategory.icon = req.body.icon;
@@ -32,7 +45,11 @@ router.post("/", async (req, res) => {
     res.send({ success: true });
   }
 });
-router.delete("/:id", (req, res) => {
+router.delete("/:id", async (req, res) => {
+  const authResult = await auth(req, res, true);
+  if (!authResult)
+    return res.status(500).json({ success: false, err: "auth err" });
+
   Category.findByIdAndRemove(req.params.id)
     .then((catfind) => {
       if (catfind) {
@@ -45,6 +62,10 @@ router.delete("/:id", (req, res) => {
 });
 
 router.put("/:id", async (req, res) => {
+  const authResult = await auth(req, res, true);
+  if (!authResult)
+    return res.status(500).json({ success: false, err: "auth err" });
+
   const category = await Category.findByIdAndUpdate(
     req.params.id,
     {
